requirements.txt
flask==3.0.2
flask-cors==4.0.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.3.2
pymongo==4.6.1
python-dotenv==1.0.1
xgboost==2.0.3 